requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
astroid==3.3.5
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
click==8.1.7
dill==0.3.9
distro==1.9.0
dj-database-url==2.3.0
Django==4.2.16
django-heroku==0.3.1
django-stubs==5.1.1
django-stubs-ext==5.1.1
exceptiongroup==1.2.2
flake8==7.1.1
flake8-bugbear==24.10.31
flake8-django==1.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
isort==5.13.2
jiter==0.7.1
mccabe==0.7.0
openai==1.54.4
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
pylint==3.3.1
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-decouple==3.8
sniffio==1.3.1
sqlparse==0.5.1
tomli==2.1.0
tomlkit==0.13.2
tqdm==4.67.0
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
uvicorn==0.32.1
whitenoise==6.8.2
